refactor(types): Splitting up types into more specific files
