+++ packages/core/src/report-components/report-component-data.ts
@@ -1,4 +1,4 @@
-import { TaskMetaData, ReportComponentType } from '../types';
+import { ReportComponentType, TaskMetaData } from '../types/tasks';
 
 export abstract class ReportComponentData {
   constructor(public meta: TaskMetaData, public componentType: ReportComponentType) {}
